Change bandwidth type from int to long.
It's required to change bandwidth type from int to long
to have a consistent type in different DTO's.

Closes #928

+++ services/src/messaging/src/main/java/org/openkilda/messaging/model/BidirectionalFlow.java
@@ -30,7 +30,7 @@ public class BidirectionalFlow {
     // private String lastUpdated;
 
     @JsonProperty("bandwidth")
-    private int bandwidth;
+    private long bandwidth;
     @JsonProperty("ignore_bandwidth")
     private boolean ignoreBandwidth;
 
@@ -65,7 +65,7 @@ public class BidirectionalFlow {
     @JsonCreator
     public BidirectionalFlow(
             @JsonProperty("flow_id")  String flowId,
-            @JsonProperty("bandwidth") int bandwidth,
+            @JsonProperty("bandwidth") long bandwidth,
             @JsonProperty("ignore_bandwidth") boolean ignoreBandwidth,
             @JsonProperty("cookie") long cookie,
             @JsonProperty("description") String description,
