+++ services/wfm/src/test/java/org/openkilda/wfm/topology/flow/PathComputerMock.java
@@ -40,7 +40,7 @@ public class PathComputerMock implements PathComputer {
     }
 
     @Override
-    public AvailableNetwork getAvailableNetwork(boolean ignoreBandwidth, int requestedBandwidth) {
+    public AvailableNetwork getAvailableNetwork(boolean ignoreBandwidth, long requestedBandwidth) {
         return new MockedAvailableNetwork();
     }
